Updates date format to abide by setting for US formatting.
