+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ep_adminpads3",
-  "version": "1.0.3",
+  "version": "1.1.3",
   "description": "Etherpad plugin to list and delete pads in /admin.",
   "homepage": "https://github.com/RpNation/ep_adminpads3",
   "bugs": {
